+++ assets/tests_with_conditional_includes/src/nested_extra.h
@@ -1,0 +1,13 @@
+/* =========================================================================
+    Ceedling - Test-Centered Build System for C
+    ThrowTheSwitch.org
+    Copyright (c) 2010-26 Mike Karlesky, Mark VanderVoord, & Greg Williams
+    SPDX-License-Identifier: MIT
+========================================================================= */
+
+#ifndef NESTED_EXTRA_H
+#define NESTED_EXTRA_H
+
+#define NESTED_MACRO (17)
+
+#endif // NESTED_EXTRA_H