+++ assets/tests_with_conditional_includes/src/project_flag_extra.h
@@ -1,0 +1,13 @@
+/* =========================================================================
+    Ceedling - Test-Centered Build System for C
+    ThrowTheSwitch.org
+    Copyright (c) 2010-26 Mike Karlesky, Mark VanderVoord, & Greg Williams
+    SPDX-License-Identifier: MIT
+========================================================================= */
+
+#ifndef PROJECT_FLAG_EXTRA_H
+#define PROJECT_FLAG_EXTRA_H
+
+#define PROJECT_FLAG_MACRO (13)
+
+#endif // PROJECT_FLAG_EXTRA_H